Build FLIPS with additional flags so it doesn't require DLL
FLIPS's Makefile seems to not include the necessary flags to make the resulting windows EXE actually work unless `libgcc_s_dw2-1.dll` is on PATH for some reason. I recompiled it with the necessary flags and I believe it should work out of the box now. 

It's also freshly pulled from its git repo instead of taken from that one comment on its tool page on SMWC, which is probably a good thing.